AICare-baselines/requirements.txt
lightning
wandb
hydra-core
torchmetrics
numpy
pandas
openpyxl
scikit-learn
catboost
xgboost
